Eilmer3: finish the implementation of write_at_step and stop writing 9999 files.
e3post.py can now cope with solution files having tindx=xxxx.
Also, changed the behaviour of the main code to use a simply incremented
value of tindx at the end of the simulation.  This means that there won't be
a set of tindx=9999 files left behind.  e3post copes with this by searching for
the last flow solution frame if the user specifies --tindx="9999" or "last".

nenzfr has also been changed to look for flow data files with tindx=0001