onedval: Added user-defined average property feature and increased robustness.
In this commit, I've added a feature that allows a user to define
a function which is applied to every cell and summed over all cells.
This allows one to compute quantities useful when calculating things like
mixing efficiency and combustion efficiency. I decided not to implement
the calculation of those values directly because various definitions
of those values exist in the literature (and they change based on global
equivalence ratio). So, it seemed more sensible to let the user define
these calculations themselves. That way, there is not ambiguity about how
the value was computed when they come to report their results.

Also in this commit, I made available a number of extra quantities
when computing flux-conserved quantities such as:
R, gamma, Cp, Cv, a, h, h0

I also increased the robustness of the flux-conserved quantities
calculation. A statically defined value for N_RETRIES (in prop_avg.py)
controls the number of times the optimiser will be retried with
the latest and greatest guess if it has failed to converge.