onedval: Added configuration option 'skip_bad_slices' (default: True)
Occasionally, there are some bad slices of data. I'm not too
sure if it is dodginess on Tecplot's part when preparing
slices, or dodginess in my code. In any case, onedval
will now report a WARNING when a bad slice is encountered
but continue on processing slices of data. If the flag
'skip_bad_slices' is set to False then that might be
considered a strict mode: the program will complain about
the first bad slice it finds and then quit.