lib/gas: Fixed two bugs in related to using separate vibrational modes.
BUG 1. The name constructor to assemble the various modes
was using the string find() method. This would incorrectly
identify O-electronic as being part of NO-vibration because
"O-electronic" is a substring of "NO-electronic". Instead,
we now use the string compare() method to test for identity
with a given string.

BUG 2. In my recent re-work of the e-vector, I had not considered
the case where mass fractions in a mode are set to zero. It
becomes difficult to back out temperature when the amount of
mixture-fraction weighted energy is 0.0. In this case, we know
there is no species existing in the mode, so a temperature value
makes little physical sense. (What is the temperature of non-existent
NO molecules, for example?) What we do then is patch the result
with the local translational temperature. This should be good
for most cases. For some kinetics schemes, we actually have models
about how much energy (or equivalently, temperature) molecules should
have when they form/appear. So, for these schemes, we'll have a good
number and not need to patch with the translational temperature.