Added pieces for PartiallyCatalyticWBC.
I have begun to implement a partially catalytic wall boundary
condition based on the architecture of other catalytic wall BCs.
The actual functions to compute the BC (in bc_catalytic.cxx)
need some work but the pieces to allow the BC to run are in
place, including re-numbering of the catalytic BC types in
bc_defs (non-catalytic and shock-fitting were both '21'). When
defined in the user's python script, the partially catalytic BC
takes a lua file with the desired finite rate reactions as its
input. At present, recombination reactions only, no adsorption
or desorption to the surface (maybe later?).