lib/gas: Fixed the initialisation of sp_idx; cause of bug in Rev. 943
When moving to the new e-vector formulation, I had dropped
the initialisation of the sp_idx_ member vector in
the thermal-energy-modes class. I had dropped this because
it was only used in the computation of massf_mode which
was not longer required.

However, I decided to use the sp_idx_ vector when making
a decision about what to do in the case of vanishing
mass fraction in a given thermal mode. So the bug was that
I used this vector but it was not populated.

The Giordano case passes in this commit.