Eilmer3: simplified the code for reconstruction of thermo data at cell faces.
Moved the selection of the thermo interpolator back into one_d_interp.cxx
so that we can put some of the redundantly computed intermediate values
into module-level static variables.  This will come in a subsequent commit.

Sorry Rowan.  I replaced your pointer-based selection of the thermo
interpolator with a switch statement in the interpolation code because
I find the switch statement easier to think about.  This also allows me
to throw out quite a few files.  Fewer files to maintain is good (for me).

This commit should not change the behaviour of the code although a small bug
was weeded out for a particular corner case of failed reconstruction.