lib/gas/build: reinstated the -m32 compile flag for cea2
Rowan reminded me that cea2 used to behave strangely without this flag.
There is now a makefile in the cea2 source area so you can go and fiddle
in that directory if the compile from this directory fails.
Such a failure, however, should no longer stop the overall build process.