lib/gas: Hooked up Knab_et_al chemistry-coupling model.
The Knab_et_al model for chemistry-vibration coupling
is now hooked up and available from the user input.
The user can select "Knab_et_al" to compute nonequilibrium
reaction rates, and they may select "Knab_et_al" as
a model for the amount of vibrational energy removed
or gained during destruction/production of vibrational species.

The Knab et al. model is an extension of the Treanor-Marrone
model. The extension introduces a parameter alpha. When alpha
is set to 1.0, the Knab model is identical to the Treanor-Marrone
model. This assertion has been checked out by running the
poshax3 Treanor-Marrone test case with the Knab et al model.