lib/gas/kinetics: added Abe's electron-rotation thermal relaxation model.
This is a simple model where the ER cross-section is assumed to be equal to
the ET (electron-translation) exchange cross-section multiplied by some factor
(e.g. 10 for N2 and O2, 100 for NO).  An example of its use is given in
examples/poshax3/FireII/1634s/air-VT-ET-ER.lua.

Also in this commit the lua input files for thermal relaxation in the FireII
poshax3 example have been renamed to reflect the naming convention used internally
by the code.