Starting my new implementation of an ablating BC in eilmer3.
There is an ablating BC for eilmer3 that was created by Dan Potter
a few years back in an attempt to solve problems like the Kaattari
test case. But, this is an approximation and needs more development
to be able to solve ablation problems for aeroshell-type problems.

I have begun to mess around with the bc_ablating.cxx and .hh files,
removing things that aren't needed and adding bits and pieces. Don't
use these files through the development stages; this is mainly for
others to monitor what I'm doing and pick up my mistakes!

The files that Dan created have, for now, been renamed bc_ablating_old
.cxx/.hh and have been left in case I delete anything I shouldn't have.

Watch this space!