eilmer3: Upgrade to BCs to allow setting of fluxes directly. Stage 2 of 3.
In this commit, all the house keeping code is in place so that I can
split the user-defined convective flux and user-defined viscous flux
in the next commit. The boundary conditions now carry two new flags:

sets_conv_flux_flag :: if true, the b.c. will set the convective flux directly
sets_visc_flux_flag :: if true, the b.c. will set the viscous flux directly

Although these flags were put in for the user-defined b.c., I envisage
that they will be useful for other C++-implemented boundary conditions
where it is more natural to specify the boundary condition by setting
the flux.

The old flag use_udf_flux has now been removed as it was ambiguous
in light of the new changes. The examples/documentation will need to
be updated.

NOTE: The viscous flux b.c. override it NOT handled in the 3D case
at this stage. I will ask PJ to look at this because I got confused
with the indexing across interfaces.... it didn't seem to match what
I am used to in 2D.