eilmer3: Upgrade to BCs to allow setting of fluxes directly. Stage 3 of 3.
In this commit, the interface for the user-defined boundary conditions has
been updated to use the new function names we decided on. We also have
the facility to set the convective fluxes and viscous fluxes separately.
The functions which the user may write are:

ghost_cell() :: Returns values for the two ghost cells at a boundary.
                These values are used by the reconstructor.

convective_flux() :: Returns values for the convective flux of conserved
                     quantities at an interface that lies on a boundary.
                     When present, these fluxes are used instead of the
                     flux calculator values. This function is only applied
                     when the b.c. option "sets_conv_flux" is true.

interface() :: Returns the properties of flow primitive values at the mid-point of
               an interface that lies on a boundary. This is used to compute the
               viscous derivatives, and in turn, the viscous fluxes.

viscous_flux() :: Returns values for the viscous flux of conserved quantities
                  at an interface that lies on a boundary. When present, these
                  fluxes are used instead of those computed from the viscous
                  derivatives. This function is only applied when the b.c. option
                  "sets_visc_flux" is true.

Also in this commit, the examples "cone20-udf" and "iris-diaphragm" have
been updated to reflect the new changes.

This completes the changes to the code. In a subsequent commit, I will
get the documentation up to date.