examples/poshax3: added an example of modelling shock heated H2-He mixtures
Also in this commit, the tube_width input parameter in the [radiation] section
of poshax3 input files has been renamed to path_length.  This will hopefully make
it more clear that this is length used for line-of-sight calculation, and can either
be used for a shock tube (the path length is the tube width) or a subscale model
calculation (the path length is the cylinder length, for example).