lib/gas: fixed bug in parsing of energy exchange mechanisms
The previous implementation assumed there was at least one entry
in the energy exchange input file for each of the nonequilium
thermal modes.  While this should usually be the case for a
complete energy exchange model (i.e. if you want to describe
a set of thermal modes by a separate temperature, you should
probably have some energy transfer mechanisms into that mode),
for debugging and sensitivity analyses one may wish to describe
a thermal mode without any energy transfer mechanisms.  This is
analogous to having a chemical species that doesn't participate
in any of the chemical reactions given, and should be allowed.
The changes in this commit allow for this scenario.