eilmer3: Added extra info to be passed for a UDF interface() function.
When the user provieds an interface() as part of a user-defined b.c.,
there are extra arguments now supplied:
  o dt      : the global dt
  o t_level : the current time-level within update
  o fs      : the flow state at the interface centre

Additionally, there are three new service function which return
information about interfaces as an analogue to the sample_flow()
function for cell information:
  1. sample_i_face()
  2. sample_j_face()
  3. sample_k_face()
These are untested at the moment.

Also, a new entry was added to the global_data struct.
That is: size_t t_level <-- Used to record the time-level
within an explicit increment.