examples/photaura: Added gnuplot script for the OH-validation test case
Added script to plot comparisons of the spectra between Photaura and
SPARTAN, with the same input data in both programs. Also added the
spectra from SPARTAN for this comparison.

The radiator_library.py file has been changed as well, i.e., the number
of levels reduced in the Einstein coefficients matrix, in order to have
the same input as in SPARTAN (this has been done to have the best
comparison possible with the SPARTAN code, and can be later reverted to
include more levels, such as in LIFBASE).