eilmer3/utilities: A small script to add mass fractions to LUT-gas simulations.
This script should be considered a demo script as to how to
use the eilmer python modules to do a customised post-processing
of flow data. In this customisation, an equilibrium air simulation
(via look-up table) is expanded into its component mass fractions
and these are written to Tecplot output. It's not very general
as it has some very case-specific options hard-coded into the
script.

Our short-term plan is to add this particular capability to the
e3post program in a generalised way. This script can remain in the
repository though as demo of how to pick-up flow solutions and
interact with them for customised post-processing.