Eilmer3 house-keeping: simplify the creation of boundary-condition objects.
These changes are limited to the C++ code and move the reading of the .config
dictionary (to bet boundary condition parameters) closer to within create_BC().
This change reduces the number of parameters which need to be passed to that
function and should make the addition of new BCs easier.

This version is a check point for me because I have to go off and do other things
for a little while.  It compiles but I'm sure that it's incomplete.