Eilmer3: simplify the creation of boundary-condition objects -- part 2.
Parameters for boundary conditions have now been divided into three sets.
(1) Those that are common to all BCs and need to be set for all classes of BC.
(2) Those that are common to all BCs and have the same values for many BCs.
Such parameters are used by other parts of the code (typically the
reconstruction and flux calculation functions, I think) and need to be accessible
via pointers to the base class (BoundaryCondition) objects.  These are kept
within the base class but may set to other than default values by the
derived BC classes.
(3) Those that are specific to only one BC.  These are accommodated within the
derived BC class.

It should now be easier to add new BoundaryCondition classes because we don't need
to carry the whole bag of parameters around as be create the BoundaryCondition
objects at initialization.

There is one (temporary) hitch.  Most of the test cases run, however, the UDF test
cases are broken at the moment.  There will be another commit...