lib/radiation: improvements to 'make test'
Spectra from the various models are saved to disk.
Tolerable difference to expected results is increased to 1%.
Spradian gives slightly different results when comping with
gfortran4.8 compared to gfortran4.2.  Also I had to modify the
spradian source to get it to compile with 4.8, there were some
errors relating to 4 or 8 bit precision of real numbers.