eilmer: Subsonic b.c. (pressure formulation) debugged.
I finally found the time to debug the implementation
of a pressure-based formulation of the subsonic b.c.
The main bug was in the thermo routine which evaluates
the gas state from a supplied pressure and entropy.
I had an erroneous call to mixture_enthalpy() instead
of mixture_entropy() deep in the secant iterations.