Eilmer3: added --binary-format option to e3post.py.
This option writes the VTK-XML files using binary data for the floating point
quantities.  The result is a mixture of text and binary that is considerably
smaller than the default all text (ascii) format.
For example the cone20-simple case the .vtu files for the final time reduce from
158453 and 468877 bytes to 54905 and 160159 bytes, respectively.  This is nearly
a factor of 3 reduction.