Eilmer3 examples: add Mohammadian's convex ramp experiment.
Following on from the concave-ramp, this one seems to be an even better example
of laminar flow with a simple and well-defined geometry.
This time there is dimensional heat-flux data and strong favourable pressure
gradient in which the boundary layer develops.
The experiment was done in Imperial College's gun tunnel and Mohammandian finished
his thesis in 1970.  Some data is published in JFM, Vol 55, pp165-175 (1972).