onedval: Added more capability to the user-defined calculation of integrated quantities.
Mat Bricalli requested the capability to compute a distortion
parameter using onedval. Distortion is an integrated quantity of
sorts but requires the average momentum flux across the plane.
The "distortion" is a measure of how much the local momentum flux
at any location deviates from the average momentum flux. The code
required a small bit of re-work so that the user-defined function
had access to the total fluxes across the plane of mass, momentum
and energy.