Eliminated the names son and u as aliases for sonic velocity and internal energy.
These names were in use because of the notation used by the CEA code and I thought
that it would be good to have names consistent with that code, however, the rest
of our code base used a and e for those quantities.  Having both sets of names in
code that got used beyond the original intent eventually led to surprising behaviour
that was noticed by Chris James.  We'll try to use the names a and e consistently
from now on.