L1d3: eliminated a couple of warnings emitted by gcc 4.8.1.
This exercise in yak shaving was an attempt to get past the dodgy optimized code
from gcc 4.6.3 on my Ubuntu box.  There, L1d3 segfaults on Chris James' x2 example
if the default build is used, but works fine if the build is done for_gnu_debug.
Also, on a Fedora 19 box with gcc 4.8.1, there appears to be no problem with the
default build at all.

I guess that there's something in the source code that upsets the 4.6.3 compiler
but I have no idea what it is.  For the moment, users of L1d3 are advised to
build with gcc 4.8.1.