Eilmer3: generalize the use (or not) of ghost-cell data in the convective fluxes.
This allows to set fluxes in some boundary conditions and then not have to bother
with ghost-cell data for those boundaries.  Presently such boundaries are the
User-Defined-Function BC when sets_conv_flux flag is set and also Andrew Pastrello's
shock-fitting BC.  No doubt, there will be more in time.  Elise's ablation BC comes
to mind.

The main changes are to more selectively use the ghost-cell data, checking if it
is actually available for use in the reconstruction process that occurs before
application of a flux calculator.

As part of the clean-up, some one-sided interpolation functions were moved to
Andrew's shock-fitting BC module because that's the only place they're used.
Also, the 4-point one-sided reconstruction function was deleted because it was
replaced with a slightly simpler (but lower-order, I presume) arrangement for
one-sided reconstruction that reused the 4-point symmetric code.  It may be worth
reinstating the higher-order reconstruction if it works much better for little
extra computational effort.