Eilmer3: added a revision string that gets printed at program start.
The intention is for people to be able to report accurately the version of the
code that they happen to be running when things go wrong.

At compile time, the Mercurial revision identity is obtained via a shell call
from within the makefile.  The resulting string is then substituted into the
kernel source code before compilation.  The string is printed when starting the
main simulation code as well as e3prep.py and e3post.py.

Of course, this works on my Linux machine just fine.
Please tell me ASAP if it fails in your environment.