Eilmer3: completed work on --binary-format option for e3post.py.
This option writes the VTK-XML files using binary format for the data arrays.
The result is a mixture of text header and binary data that is considerably
smaller than the default text (ascii) format.
For example the cone20-simple case the .vtu files for the final time now reduce
from 158453 and 468877 bytes to 53636 and 153887 bytes, respectively.

This finished the work started in rev 1074 (about 2 weeks ago) by including the
cell connectivity and type data.