Eilmer3 convex ramp case: revised to accept his polynomial curve description.
My original working was trying to be too clever at reconstructing his polynomial
to allow the aft part of the ramp to have zero slope.  The JFM paper didn't say
that it had zero slope, and the photo surely looks like it has the ramp sloping
down (away from the free-stream) for the final straight section.