eilmer: SuperBlock2D now accepts a 'psurf' or 'grid' just like Block2D
Daniel Oberg provided me with a Gridpro grid of the JAXA
nozzle. It had two blocks that were approximately 15 times
larger than the average block size, so this was going to
dominate the run time. A neat solution appeared to be to
use the SuperBlock object to split up the large blocks,
however, they would not accept a grid object. In this update,
the SuperBlock2D constructor is able to accept a previously
defined StructuredGrid object.