lib/gas: added electrical conductivity to the Gas_data structure
This parameter is refered to in the codes as 'sigma' and has SI units
of Siemens/meter (S/m).

Currently only the thermal nonequilibrium gas-models compute a non-zero
value for sigma, and only when the Gupta-Yos transport model is used.
This is because collision integrals are required for its calculation.
Currently a simple first order approximation (to Chapman-Enskog theory)
for sigma is implemented, from NASA TM No. RAD-TM-63-7 by Yos.

Todo: add sigma to the IO functionality.