lib/plot: Removed the active_ps_path logic from the postscript plotting kernel.
Occasionally, the line-clipping functions would mess with the postscript paths
in a way that caused the plotter to drag its virtual pen across the field of view.
The easiest fix has been to remove the small optimisation offered by keeping paths
active and just stroking all of the segments individually.

For Hadas' x2 plot, the penalty has been an increase in postscript file size from
29MB to 30MB, however, on conversion to PDF (using epstopdf) the difference is
about 15kB out for a 3.9MB file.  Negligible.