eilmer3: added reflection at walls for the Monte-Carlo radiation transport model
Rays incident on walls are now diffusively reflected, as outlined in the text
of Modest.
For the standard absorption model, the photon is absorbed by the wall element
when a randomly generated number is less than the surface emissivity.
For the partitioned energy absorption model, the fraction of energy absorbed by
the wall element is epsilon, and the photon continues post-reflection with the
remaining energy.

Currently photons are re-emitted from the wall element centroids, and epsilon is
set to 1.0 for all boundaries.  The test cases in examples/eilmer3/2D/radiating-cylinder
all pass on my machine.

Todo: set real epsilon values, test with epislon < 1.