Out with the old, in with the new Ablating BC.
Taking out all of the old bits so that we can start again on the
Ablating BC. The convective and viscous fluxes are copied from
bc.cxx and bc_fixed_t.cxx, respectively, and the f and Jac functions
required by the zero solver are in a very basic state. Watch this
space!