nenzfr: Delegate actual block-marching work to e3march.
In theory, this should have been a simple swap, however, there is a bug.
The Mach 4 nozzle (with equilibrium chemistry) seems to show that the inflow
temperature is 300K (or there about) rather than the correct value of 3788K.
Other properties such as pressure and velocity seem to be set correctly.

There will be another commit...