e3march.py: Limit the time step set at the start of each time-integration.
The strategy of carrying the time-step at the end of a time-integration
for one slice of blocks onto the next set of blocks can run into problems
if the wall geometry changes suddenly, as it does for the circular-arc bump.
If you have such a situation, the option --max-dt will provide a way to limit
the starting time step for each time integration and keep things stable.

Also, moved back to using Chris' configuration dictionary because we may have
to introduce more controlling parameters in the future and the argument list to
the main function was growing too long for my liking.