Eilmer3: a little tidy-up of code for e3march.py (no functional change).
There is still a bug that prevents the code from running the nenzfr Mach 4 case
but it's shared between the ConfigParser behaviours.  The Python ConfigParser
writes parameter names in lower case, irrespective of the case of the original.
Rowan's ini-file parser, however, is case sensitive.  The fix for this mismatch
will be in a subsequent commit.