nenzfr: fixed bug in Mach 4 nozzle test scripts.
This bug showed up only when running two, supposedly separate tests in the one
directory.  The main simulation code was doing fine but the second test was
failing because it was looking at an accumulation of result flags rather than
starting with a clean slate.  (And, yes, used to have a slate for practising my
writing when I was a young lad.)