Eilmer3: added gdata.dt_max to allow the user to directly limit the time step.
This parameter was added to the .control file to see if it could be used to
stabilize Elise's simulation with strong source-terms for the porous media.

Along the way, the User Guide needed a little fixing because some of the Hayabusa
simulation files had moved.  Also, updated the usage message for e3post.py.