lib/radiation: a working implementation of TOPBase photoionization cross-sections
The bounding data points for a given frequency are found via a bisection method.
Also in this commit, the example comparing NIST and TOPBase in Argon/pT has been
expanded to include a plot of the comparison.  TOPBase is seen to predict much
stronger radiation due to an intense line in the VUV.