lib/gas: Implemented temperature limits on evaluation of reaction rate coefficients.
The reactions are no longer frozen by the finite-rate chemistry
module if the upper or lower limits are exceeded. Instead,
the limits are used to enforce a limit on temperature values
used in the evaluation of the rate coefficients.

For nonequlibrium rates, I have implemented the following:
1. The limits are used to limit the evaluation of BASE reaction
   rate coefficient (typically the Arrhenius rate coefficient)
2. The limits are NOT used to compute the nonequilibrium factor
   applied to the base rate.

Also in this commit, I have removed the building of the Lua gas module.
This was not being used by anyone and I no longer wanted to maintain
it. The associated files have been removed and references to its
build are gone from the makefile.