app/poshax3: added option to apply the 'u dpe / dx' source term to the electronic energy equation
This source term was proposed by Lee (AIAA-84-1729, Eq. 34) to account for "the
work done on electrons by an electric field induced by the electron pressure
gradient".  It is based on the approximation that the electric field due to
charge separation is proportional to the electron pressure gradient.

The validity of this term is unclear.  In the literature, half the papers seem
to include it while the other half don't.  Possible justifications for its
omission are (a) charge neutrality is assumed, and (b) the electric field is zero
due to the low magnetic Reynolds number approximation.  In any case, it can now
be included in a crude way (backward differencing for the derivative) in poshax3,
so it is possible to check its influence on the solution.  For a Mach 16.5 argon
shock tube case, its inclusion reduced the relaxation time by about 10 percent.

See the poshax_spec.cfg file for how to turn it on and off.