Reinstated the look-up-table gas models built by CEA2.
Updated the Eilmer3 user guide to be consistent with the current naming.
Have also added the CEA2 gas specifications to the small library inside
cea2_gas.py so that we can call up the exact corresponding models to those
in the look-up tables.  Also, have added the shell files used to generate
the tables so that all of the parameters used have made it into the repo.

Finally, removed the old Lua version of the table building code and its
input files.