First working reacting wall test using udf BC.
This example is the first implementation of a reacting wall to
model ablation in eilmer3, by way of a user-defined boundary
condition. The fluxes are determined by the mass flow rates of
species produced and depleted by gas-surface interactions at
the wall. Here, a solid carbon surface reacts with an oxygen
flowfield (n90 geometry and flow conditions). The image shows
CO produced at the wall after about 30000 steps.

This is not at all validated and certainly a work in progress,
but it runs and reactions happen at the wall and I'm excited!