app/eilmer3: implemented electric field work term for free electron energy
This term accounts for the work done on free electrons due to the electric
field induced by ambipolar charged particle diffusion.  This implementation
uses the approximation that the work is equal to vec(u) . div(p_e) as
derived by Lee in AIAA-84-1729 [see equations 30, 36 and 37].

To implement this model the derivatives of the electron pressure have to
be stored and calculated.