NENZFr: Add Mach 7 nozzle coordinates (plus small mods to input.template)
Added the coordinates of the new T4 Mach 7 nozzle. Also modified the
nozzle.input.template file such that a block-marched simulation is always
started with the entire flow domain initialised to the throat conditions.
For transient simulations, the setting is left as-is, the flow domain is
initialised to the initial conditions in the T4 test section.