lib/gas: Brought over Knab model modifications from libgas2.
The expressions in the Knab model suffer from problems
of representation in machine precision at low temperatures.
What this means is that denominators become zero in some
expressions, and the division by zero propagates. The solution
I used in libgas2 and have brought across here is to treat
the low temperature cases as being close to thermal equilibrium.

1. For the modification of the reaction rate, the nonequilibrium
   factor is simply set to 1.0. (So no modification).

2. For computing the average vibrational energy of appearing
   and dissappearing molecules at low temperatures, it is assumed
   they have an energy level corresponding to the translational
   temperature.