Separate time delay and increment for diffusion terms.
Viscous effects can be introduced after some time delay and
ramped in an attempt to maintain stability, and diffusion
was either coupled to this or introduced fully at the start
of run-time. Diffusion now has its own time delay and
increment that are global data variables, so that if desired,
it can be decoupled and introduced separately to viscous
effects. At present, if you want them to be introduced in the
same way, give the delay and increment the same values.

I'm hoping that by gently including diffusion, we can get
over some stability issues in problems such as aeroshells.