eilmer3: Added infrastructure for conjugate heat transfer modelling.
This commit compiles cleanly. However, the functionality of
conjugate heat transfer modelling is not completely in place.
New boundary condition files bc_conjugate_ht.hh/.cxx have been
added but these are not hooked up yet (and the makefile does
not compile them). The rest of the implementation can happed
when I sit with Jared Clifford to bring in his pieces.