Eilmer3: start clean-up of code relating to VISCOUS_TIME_LIMIT_MODEL.
Some more needs to be done here because the model for diffusion is currently not
functional and the 2D application uses edge lengths that may go to zero for cells
on boundaries that have zero length.  Such cells are OK so long as the (x,y)-plane
area does not go to zero.  Should use iLength anf jLength as done for the 3D case.