lib/nm: Additional functions for no_fuss_linear_algebra.
I have added Jared Clifford's addition to the mini linear
algebra library. Jared's additions are to help with solving
a system of equations related to heat conduction in a solid wall.

While I was in the file, I tidied up some other functions.
For example there were two versions of the gaussian_elimination:
1. as a method in Valmatrix class, and
2. as a function external to the class.
I prefer the function because we don't need to know anything
about how a Valmatrix is stored in order to do an elimination.
I'm not sure of the history of the method version. Presently,
I've commented it out. If it breaks some code, it will be an easy
fix to use the function version.