eilmer3/twc: Further work on the conjugate heat transfer implementation.
I have done some preliminary testing. I fixed up a segfault related
to the fact that I did not insert the new boundary condition name
into the map in init.cxx.

The connection between the flow solver and the wall model seems
to be behaving, in as much as data is going back and forth.
The implementation now needs some rigorous testing for correctness.