gas/eilmer3: Added flag to indicate if gas model is compatible with finite-rate chemistry module.
When the thermal behaviourmodel is set, we can make a determination about whether or
not the thermodynamics of teh gas model are compatible with the finite-rate chemistry module.
The default assumption is that gas models are NOT compatible (set in the base constructor
Gas_model). In the case that the gas model IS compatible, the flag "good_for_reactions" is
set to true when setting the Thermal_behaviour_model in the constructor for a Composite_gas_model.

The calling code may test if a gas model is compatible with finite-rate chemical reactions
by using the method:

gmodel->good_for_reactions()

it return true if it is compatible, and false otherwise.