eilmer3: MPI version of conjugate heat transfer modelling working.
The MPI version now works when the conjugate heat transfer is active.
I've tested the problem of an isothermal system over two blocks
using two MPI_ranks and things behave as expected.

WARNING: The e3shared.exe that is generated when compiling
the MPI executable does NOT WORK when the conjugate heat
transfer is active. I need to fix this soon.