eilmer3: Explicit request for pT interpolator in Giordano test case.
An earlier bug in rev 1309 (and back) had the Giordano test
case using the pT interpolator when we expected the rhoe interpolator.
Rev 1310 fixed this buggy selection and so the Giordano test
failed because the rhoe interpolator was being used (correctly)
but the reference solution was built on the pT solution.

Instead of updating the reference solution, I decided to
explicitly request the pT interpolator. So now the test
passes again and we have a test of the pT interpolator in
the suite of short tests.