Eilmer3: update the 2D signal speed calculation to use iLength, jLength
...rather then the cell edge lengths.  This makes the calculation more
consistent with the 3D version and should remove one more catch when
running simulations on grids that have one edge collapsed to a point.
An example of such a grid is used in the spherical implosion simulations
done for Vince and Dale.

All of the short automated tests pass so it seems that this change was only
an infinitesimal perturbation.  Good.