Eilmer3: added option to change flow orientation between connected blocks.
After the data is exchanged betwwen connected blocks, we may optionally reorient
the vector flows quantities in the ghost cells.  These quantities are presently
velocity and magnetic field.

This option allows us to connect block boundaries that have differing
orientations, as opposed to being directly face-to-face.  It is a useful
capability if one is trying to simulate flows in circular geometries
but with only a sector of the full geometry actually being in the calculation.

At input, the user has to manually specify block connections with a normal
and tangent vector giving the nominal orientation of each connecting face.
See the examples/eilmer3/3D/simple_ramp for how this might be done.