Nozzle optimiser: Added example for the optimisation of the Mach 4 nozzle contour
Brought scripts for the nozzle optimiser up to date with the current nenzfr/eilmer
scripts. This example demonstrates the optimisation of the contour of the T4
Mach 4 nozzle. The boundary layers in the nozzle were assumed to be laminar and
the flow through the nozzle to be in chemical equilibrium.