Start to make libgas_gas a little more general.
We want Rowan's libgas models to be available to the cfpylib/gasdyn functions.
Peter Blyton had made the original libgas_gas.py module to make his REFPROP
and Bender high-density models available, however, we want more of the libgas
models available.  Specifically, we want the look-up-table to be available in
ESTCj for cases where the cea_gas model fails.