Fill in the implementations for a couple of the service functions in the LUT gas.
These functions are needed for the libgas_gas module.
For the moment, we have put in expressions based on an approximate ideal gas
model, however, we should more completely implement the entropy look-up.
This will, however, require us to store entropy in the look-up-table.
We should make this an optional feature so that calculations that don't require
it can still work with the current tables.