LUT: continued work on the inclusion of entropy in the look-up-table.
Note that this is still (subtly) borken for low-temperature air, at least.

Also, note that the extrapolation limits for the LUT were removed in this
revision.  The limits were interfering with the estimates of entropy at
room-tempeprature conditions when the table was generated with a lower
temperature limit of 500K, as is the case for the air-ions table.
Be aware that this removal of extrapolation limits may change the calculations
other people have been doing...