LUT: changed the stored variable to be Cp_hat, which is used to compute entropy.
Interpolating and extrapolating Cp_hat should be much better than entropy
directly, of course.  We should put in some notes about interpolating on
parameters of a pseudo-ideal-gas model.  This is why the LUT has worked
so well for such a long time.

Have reinstated the extrapolation limits but still need to test their effect
extrapolating to low-T conditions.  At least, old calculations should behave
exactly as they used to, so we should be able to ignore the caution in the
previous revision.