Eilmer3 house-keeping: clean up hard-coded configuation.
Move away from using the C-preprocessor to using constexpr values and normal
if statements.  Also, make more use of the bool type.

This version may have fixed up some of the code in the Baldwin-Lomax turbulence
model.  I suspect that it was (accidentally) turned off months ago when I changed
some of the flag definitions from "#define" to "const int" but left the #if #else
#endif code in place.